refactor(chain-service): define typestate for phases
