+++ nodes/node/http-client/src/lib.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use futures::{Stream, StreamExt as _, TryStreamExt as _};
 pub use lb_chain_broadcast_service::BlockInfo;
-pub use lb_chain_service::{ChainServiceInfo, ChainServicePhase, CryptarchiaInfo, Slot, State};
+pub use lb_chain_service::{ChainServiceInfo, CryptarchiaInfo, PhaseTag, Slot, State};
 pub use lb_core::events::{Event, Events, TxEventPayload};
 use lb_core::{
     block::MAX_BLOCK_TRANSACTIONS_SIZE,
