+++ tests/src/cucumber/steps/manual_nodes/utils.rs
@@ -9,7 +9,7 @@ use std::{
 use cucumber::gherkin::Table;
 use futures::future::try_join_all;
 use hex::ToHex as _;
-use lb_chain_service::{ChainServiceInfo, ChainServicePhase, CryptarchiaInfo};
+use lb_chain_service::{ChainServiceInfo, CryptarchiaInfo, PhaseTag};
 use lb_core::mantle::{Utxo, ops::OpId as _, traits::GenesisTx as _};
 use lb_http_api_common::paths::CRYPTARCHIA_INFO;
 use lb_libp2p::PeerId;
@@ -1140,7 +1140,7 @@ async fn verify_online(
         let mut mode_online = false;
         match client.consensus_info().await {
             Ok(val) => {
-                if matches!(val.phase, ChainServicePhase::Following) {
+                if matches!(val.phase, PhaseTag::Following) {
                     mode_online = true;
                 }
             }
@@ -1832,7 +1832,7 @@ pub(crate) async fn get_cryptarchia_info_all_nodes(world: &CucumberWorld, step: 
         };
         match node_info.started_node.client.consensus_info().await {
             Ok(consensus) => {
-                let mode = if matches!(consensus.phase, ChainServicePhase::Following) {
+                let mode = if matches!(consensus.phase, PhaseTag::Following) {
                     "Online"
                 } else {
                     "Bootstrapping"
